done with auth. cleaning now

+++ server/api/travel/index.ts
@@ -4,6 +4,8 @@ import { Place } from './place.entity'
 import { validation_errors } from './validation-errors'
 import { myDataSource } from "../../app-data-source"
 import { DeleteResult, Repository } from 'typeorm';
+import { authGuard } from './user.controller'
+import { User } from './user.entity';
 
 const router = express.Router();
 const placeRepository: Repository<Place> = myDataSource.getRepository(Place);
@@ -21,6 +23,7 @@ function whitelist<T, K extends keyof T>(obj: T, ...keys: K[]): Pick<T, K> {
 
 // ADD
 router.post("/",
+    authGuard,
     body('title', validation_errors.title).isString().isAlphanumeric(undefined, {ignore: " "}).isLength({min:1, max: 20}),
     body('title').trim().escape(),
     body('description', validation_errors.description).isString().isAlphanumeric(undefined, {ignore: " "}).isLength({min:1, max: 60}),
@@ -35,7 +38,8 @@ router.post("/",
         const errors = validationResult(req);
         if (!errors.isEmpty())
             return res.status(400).json({ errors: errors.array() });
-        const place: Place = placeRepository.create(whitelist(req.body, "title", "description", "localisation", "score", "favorite"));
+        const user: User = res.locals.user;
+        const place: Place = placeRepository.create({...whitelist(req.body, "title", "description", "localisation", "score", "favorite"), user: user});
         const results: Place = await placeRepository.save(place);
         return res.status(201).json(results);
     }
@@ -43,6 +47,7 @@ router.post("/",
 
 // UPDATE
 router.put("/",
+    authGuard,
     body('id', validation_errors.id).isInt(),
     body('id').toInt(),
     body('title', validation_errors.title).isString().isAlphanumeric(undefined, {ignore: " "}).isLength({min:1, max: 20}),
@@ -59,21 +64,24 @@ router.put("/",
         const errors = validationResult(req);
         if (!errors.isEmpty())
             return res.status(400).json({ errors: errors.array() });
+        const user: User = res.locals.user;
         const place_received: Place = placeRepository.create(whitelist(req.body, "id", "title", "description", "localisation", "score", "favorite"));
-        const place_to_modify: Place | null = await placeRepository.findOneBy({
-            id: place_received.id
+        const place_to_modify: Place | null = await placeRepository.findOne({
+            where: {id: place_received.id},
+            relations: ['user']
         });
-        if (place_to_modify) {
-            placeRepository.merge(place_to_modify, place_received);
-            const results: Place = await placeRepository.save(place_to_modify);
-            return res.status(200).json(results);
-        }
-        return res.status(400).send({error: "Resource doesnt exist"});
+        if (!place_to_modify)
+            return res.status(400).send({error: "Resource doesnt exist"});
+        else if (place_to_modify?.user.id !== user.id)
+            return res.sendStatus(403);
+        placeRepository.merge(place_to_modify, place_received);
+        const results: Place = await placeRepository.save(place_to_modify);
+        return res.status(200).json(results);
 });
 
 // DELETE
-// the status returns from this needs to be changed
 router.delete("/:id",
+    authGuard,
     param('id', validation_errors.id).isInt(),
     param('id').toInt(),
     async (req : Request, res : Response) => {
@@ -81,23 +89,30 @@ router.delete("/:id",
         if (!errors.isEmpty())
             return res.status(400).json({ errors: errors.array() });
         const id: number = Number(req.params.id);
-        const result: DeleteResult = await placeRepository.delete(id);
-        if (result.affected)
-            return res.status(200).send({status : 'Place deleted'});
-        return res.status(404).send({status : 'Place was not deleted because it was not found'});
+        const user: User = res.locals.user;
+        const place_to_delete: Place | null = await placeRepository.findOne({
+            where: {id: id},
+            relations: ['user']
+        });
+        if (!place_to_delete)
+            return res.status(404).send({status : 'Place was not deleted because it was not found'});
+        if (place_to_delete?.user.id === user.id)
+        {
+            const result: DeleteResult = await placeRepository.delete(id);
+            if (result.affected)
+                return res.status(200).send({status : 'Place deleted'});
+        }
+        return res.sendStatus(403);
 });
 
 // GET ALL OR GET ALL FAVORITES
-// add more info if no elems? like no elems in db?
 router.get("/",
+    authGuard,
     query('favorite', validation_errors.favorite).isBoolean().optional(),
     query('favorite').toBoolean(),
     async (req : Request, res : Response) => {
         const favorite: boolean = Boolean(req.query.favorite);
-        let places: Place[] = await placeRepository.find();
-        // let test = myDataSource.getRepository(User).create({email: 'lola', password: 'haha'});
-        // await myDataSource.getRepository(User).save(test);
-        // console.log(await myDataSource.getRepository(User).findOneBy({email : 'lola'})); //--> this is how u get relations
+        let places: Place[] = await placeRepository.find({relations: ['user']});
         if (places.length === 0)
             return res.status(404).send();
         if (favorite === true)
@@ -106,8 +121,8 @@ router.get("/",
 });
 
 // SEARCH WITH A STRING IN TITLE, DESCRIPTION OR LOCALISATION AND RETURN AN ARRAY OR 404 NOT FOUND
-// add more info if no elems? like no elems in db?
 router.get("/search",
+    authGuard,
     query('keyword', validation_errors.keyword).isString().isAlphanumeric(undefined, {ignore: " "}).isLength({min:1, max: 60}),
     query('keyword').trim().escape(),
     async (req : Request, res : Response) => {
@@ -127,21 +142,18 @@ router.get("/search",
 
 // GET ALL PLACES WITH SCORE EQUAL OR LARGER THAN THE PARAMETER SCORE
 router.get("/score",
+    authGuard,
     query('score', validation_errors.score).isInt(),
     query('score').toInt(),
     async (req : Request, res : Response) => {
         const errors = validationResult(req);
         if (!errors.isEmpty())
             return res.status(400).json({ errors: errors.array() });
+        const user: User = res.locals.user;
         const score: number = Number(req.query.score);
-        const places: Place[] = await placeRepository.find();
+        const places: Place[] = await placeRepository.find({relations: ['user']});
         if (places.length > 0) {
-            const returned_places: Place[] = [];
-            places.forEach(place => {
-                if (place.score >= score)
-                    returned_places.push(place);
-                }
-            );
+            const returned_places: Place[] = places.filter(place => place.score >= score && place.user.id === user.id);
             if (returned_places.length > 0)
                 return res.status(200).json(returned_places);
         }
@@ -150,6 +162,7 @@ router.get("/score",
 
 // SEARCH WITH ID
 router.get("/:id",
+    authGuard,
     param('id', validation_errors.id).isInt(),
     param('id').toInt(),
     async (req : Request, res : Response) => {
@@ -157,18 +170,13 @@ router.get("/:id",
         if (!errors.isEmpty())
             return res.status(400).json({ errors: errors.array()});
         const id: number = Number(req.params.id);
-        const place: Place | null = await placeRepository.findOneBy({
-            id: id
+        const place: Place | null = await placeRepository.findOne({
+            where: {id: id},
+            relations: ['user']
         });
         if (place)
             return res.status(200).json(place);
         return res.status(404).send();
 });
-
-/*
-Authentication
-2fa??
-Lieu favoris, route to add and to delete
-*/
 
 export default router;
